fixing contact issues

+++ PerfectJazz/components/cmp_bullet.cpp
@@ -9,14 +9,14 @@ void BulletComponent::update(double dt) {
   }
 }
 
-BulletComponent::BulletComponent(Entity* p, float damage, float lifetime)
+BulletComponent::BulletComponent(Entity* p, uint16 damage, float lifetime)
     : Component(p), _lifetime(lifetime), _damage(damage) {}
 
-void BulletComponent::setDamage(float damage)
+void BulletComponent::setDamage(uint16 damage)
 {
     _damage = damage;
 }
 
-float BulletComponent::getDamage() const {
+uint16 BulletComponent::getDamage() const {
     return _damage;
 }
