+++ PerfectJazz/components/cmp_bullet.h
@@ -5,14 +5,14 @@
 class BulletComponent : public Component {
 protected:
   float _lifetime;
-  float _damage;
+  uint16 _damage;
 
 public:
   void update(double dt) override;
   void render() override {}
-  explicit BulletComponent(Entity* p, float damage = 50.f, float lifetime = 3.f);
+  explicit BulletComponent(Entity* p, uint16 damage = 50.f, float lifetime = 3.f);
   BulletComponent() = delete;
 
-  void setDamage(float damage);
-  float getDamage() const;  
+  void setDamage(uint16 damage);
+  uint16 getDamage() const;  
 };
